# openapi/delegated.openapi.json
{
  "openapi": "3.1.0",
  "jsonSchemaDialect": "https://json-schema.org/draft/2020-12/schema",
  "info": {
    "title": "OPAP Delegated Knowledge API",
    "version": "0.1.0-alpha.2"
  },
  "paths": {
    "/v1/query": {
      "post": {
        "operationId": "queryKnowledge",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$schema": "https://json-schema.org/draft/2020-12/schema",
                "type": "object",
                "properties": {
                  "sourceId": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 200,
                    "pattern": "^[a-zA-Z0-9][a-zA-Z0-9._:/-]*$"
                  },
                  "query": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 8192
                  },
                  "mode": {
                    "default": "search",
                    "type": "string",
                    "enum": [
                      "search",
                      "answer"
                    ]
                  },
                  "maxSources": {
                    "default": 5,
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 20
                  }
                },
                "required": [
                  "sourceId",
                  "query",
                  "mode",
                  "maxSources"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Knowledge results",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "oneOf": [
                    {
                      "type": "object",
                      "properties": {
                        "mode": {
                          "type": "string",
                          "const": "search"
                        },
                        "results": {
                          "maxItems": 20,
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "sourceId": {
                                "type": "string",
                                "minLength": 1,
                                "maxLength": 200,
                                "pattern": "^[a-zA-Z0-9][a-zA-Z0-9._:/-]*$"
                              },
                              "resourceId": {
                                "type": "string",
                                "minLength": 1
                              },
                              "title": {
                                "type": "string",
                                "minLength": 1
                              },
                              "uri": {
                                "type": "string",
                                "format": "uri"
                              },
                              "observedAt": {
                                "type": "string",
                                "format": "date-time",
                                "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
                              },
                              "excerpt": {
                                "type": "string",
                                "maxLength": 2048
                              },
                              "observationId": {
                                "type": "string",
                                "minLength": 1,
                                "maxLength": 200,
                                "pattern": "^[a-zA-Z0-9][a-zA-Z0-9._:/-]*$"
                              }
                            },
                            "required": [
                              "sourceId",
                              "resourceId",
                              "title",
                              "uri",
                              "observedAt",
                              "excerpt",
                              "observationId"
                            ],
                            "additionalProperties": false
                          }
                        }
                      },
                      "required": [
                        "mode",
                        "results"
                      ],
                      "additionalProperties": false
                    },
                    {
                      "type": "object",
                      "properties": {
                        "mode": {
                          "type": "string",
                          "const": "answer"
                        },
                        "answer": {
                          "type": "string"
                        },
                        "citations": {
                          "maxItems": 20,
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "sourceId": {
                                "type": "string",
                                "minLength": 1,
                                "maxLength": 200,
                                "pattern": "^[a-zA-Z0-9][a-zA-Z0-9._:/-]*$"
                              },
                              "resourceId": {
                                "type": "string",
                                "minLength": 1
                              },
                              "title": {
                                "type": "string",
                                "minLength": 1
                              },
                              "uri": {
                                "type": "string",
                                "format": "uri"
                              },
                              "observedAt": {
                                "type": "string",
                                "format": "date-time",
                                "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
                              },
                              "excerpt": {
                                "type": "string",
                                "maxLength": 2048
                              },
                              "observationId": {
                                "type": "string",
                                "minLength": 1,
                                "maxLength": 200,
                                "pattern": "^[a-zA-Z0-9][a-zA-Z0-9._:/-]*$"
                              }
                            },
                            "required": [
                              "sourceId",
                              "resourceId",
                              "title",
                              "uri",
                              "observedAt",
                              "excerpt",
                              "observationId"
                            ],
                            "additionalProperties": false
                          }
                        },
                        "observationId": {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 200,
                          "pattern": "^[a-zA-Z0-9][a-zA-Z0-9._:/-]*$"
                        },
                        "model": {
                          "type": "object",
                          "properties": {
                            "providerId": {
                              "type": "string",
                              "minLength": 1,
                              "maxLength": 200,
                              "pattern": "^[a-zA-Z0-9][a-zA-Z0-9._:/-]*$"
                            }
                          },
                          "required": [
                            "providerId"
                          ],
                          "additionalProperties": false
                        }
                      },
                      "required": [
                        "mode",
                        "answer",
                        "citations",
                        "observationId",
                        "model"
                      ],
                      "additionalProperties": false
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "type",
                    "title",
                    "status",
                    "requestId"
                  ],
                  "properties": {
                    "type": {
                      "type": "string",
                      "format": "uri"
                    },
                    "title": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer"
                    },
                    "requestId": {
                      "type": "string"
                    },
                    "errors": {}
                  }
                }
              }
            }
          },
          "403": {
            "description": "Policy denied",
            "content": {
              "application/problem+json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "type",
                    "title",
                    "status",
                    "requestId"
                  ],
                  "properties": {
                    "type": {
                      "type": "string",
                      "format": "uri"
                    },
                    "title": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer"
                    },
                    "requestId": {
                      "type": "string"
                    },
                    "errors": {}
                  }
                }
              }
            }
          },
          "404": {
            "description": "Source not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "type",
                    "title",
                    "status",
                    "requestId"
                  ],
                  "properties": {
                    "type": {
                      "type": "string",
                      "format": "uri"
                    },
                    "title": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer"
                    },
                    "requestId": {
                      "type": "string"
                    },
                    "errors": {}
                  }
                }
              }
            }
          },
          "429": {
            "description": "Budget limit",
            "content": {
              "application/problem+json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "type",
                    "title",
                    "status",
                    "requestId"
                  ],
                  "properties": {
                    "type": {
                      "type": "string",
                      "format": "uri"
                    },
                    "title": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer"
                    },
                    "requestId": {
                      "type": "string"
                    },
                    "errors": {}
                  }
                }
              }
            }
          },
          "502": {
            "description": "Invalid source result",
            "content": {
              "application/problem+json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "type",
                    "title",
                    "status",
                    "requestId"
                  ],
                  "properties": {
                    "type": {
                      "type": "string",
                      "format": "uri"
                    },
                    "title": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer"
                    },
                    "requestId": {
                      "type": "string"
                    },
                    "errors": {}
                  }
                }
              }
            }
          },
          "503": {
            "description": "Source or metering unavailable",
            "content": {
              "application/problem+json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "type",
                    "title",
                    "status",
                    "requestId"
                  ],
                  "properties": {
                    "type": {
                      "type": "string",
                      "format": "uri"
                    },
                    "title": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer"
                    },
                    "requestId": {
                      "type": "string"
                    },
                    "errors": {}
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/capabilities": {
      "get": {
        "operationId": "listCapabilities",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Capabilities"
          }
        }
      }
    },
    "/mcp": {
      "post": {
        "operationId": "mcp",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Stateless Streamable HTTP MCP response"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}